common Policy Attributes
